requirements.txt
# ============================================================
# requirements.txt — Dependencies cho SeaBank Crawler
# ============================================================

# HTTP & Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Google Sheets API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0

# Biến môi trường
python-dotenv>=1.0.0

# Browser automation (chỉ dùng khi cần — CafeF hiện dùng JSON API)
playwright>=1.40.0

# Retry & HTTP utilities (đã có trong urllib3 của requests, explicit để rõ ràng)
urllib3>=2.0.0
